+++ apps/cowswap-frontend/src/modules/limitOrders/pure/ProtocolFeeInfoBanner/index.tsx
@@ -8,6 +8,7 @@ import { BANNER_IDS } from 'common/constants/banners'
 
 const BANNER_ID = BANNER_IDS.LIMIT_ORDERS_PROTOCOL_FEE
 const CIP_74_URL = 'https://vote.cow.fi/#/proposal/0x0c70c8cd92accee872b52614b4fa10e3e3214f45c5b6857f7e88e910607a3c1d'
+const VOLUME_FEE_DOC_URL = 'https://docs.cow.fi/governance/fees#volume-fee'
 const PROTOCOL_FEE_START_DATETIME_UTC = 'November 26, 2025 at 00:00 (UTC)'
 
 export interface ProtocolFeeInfoBannerProps {
@@ -25,8 +26,8 @@ export function ProtocolFeeInfoBanner({ margin = 'auto' }: ProtocolFeeInfoBanner
       <p>
         <Trans>
           From {PROTOCOL_FEE_START_DATETIME_UTC}, and pursuant to <ExternalLink href={CIP_74_URL}>CIP-74</ExternalLink>,
-          a 2 bps (0.02%) protocol fee will apply to all executed orders, including any limit and TWAP orders executed
-          after this time, even if they were created earlier.
+          a <ExternalLink href={VOLUME_FEE_DOC_URL}>protocol fee</ExternalLink> will apply to all executed orders,
+          including any limit and TWAP orders executed after this time, even if they were created earlier.
         </Trans>
       </p>
     </DismissableInlineBanner>